tests/services/db/operations: refactored db operations tests into different files
