+++ __tests__/services/db/operations/index.ts
@@ -1,0 +1,8 @@
+// Import all operations tests to ensure Jest runs them
+import './UserOperations.test';
+import './AircraftOperations.test';
+import './MissionOperations.test';
+import './CargoTypeOperations.test';
+import './CargoItemOperations.test';
+import './FuelOperations.test';
+import './CompartmentOperations.test';
