+++ src/services/db/operations/types.ts
@@ -96,4 +96,4 @@ export interface User {
   id?: number;
   username: string;
   last_login?: string;
-} 
+}
